+++ Sources/SmileID/Classes/Networking/SmileIDService.swift
@@ -2,31 +2,48 @@ import Combine
 import Foundation
 
 public protocol SmileIDServiceable {
+    /// Returns a signature and timestamp that can be used to authenticate future requests. This is
+    /// necessary only when using the authToken and *not* using the API key.
     func authenticate(request: AuthenticationRequest) -> AnyPublisher<AuthenticationResponse, Error>
+
+    /// Used by Job Types that need to upload a file to the server. The response contains the URL
+    /// that the file should eventually be uploaded to (via upload).
     func prepUpload(request: PrepUploadRequest) -> AnyPublisher<PrepUploadResponse, Error>
+
+    /// Uploads files to S3. The URL should be the one returned by `prepUpload`.
     func upload(zip: Data, to url: String) -> AnyPublisher<UploadResponse, Error>
-    func getJobStatus<T: JobResult>(
-        request: JobStatusRequest
-    ) -> AnyPublisher<JobStatusResponse<T>, Error>
-    func getServices() -> AnyPublisher<ServicesResponse, Error>
 
     /// Query the Identity Information of an individual using their ID number from a supported ID
     /// Type. Return the personal information of the individual found in the database of the ID
     /// authority. The final result is delivered to the url provided in the request's `callbackUrl`
     /// (which is required for this request)
     /// - Requires: The `callbackUrl` must be set on the `request`
-    /// - Parameter request: The Enhanced KYC request
-    /// - Returns: A response indicating whether the request was successfully submitted or not
     func doEnhancedKycAsync(
         request: EnhancedKycRequest
     ) -> AnyPublisher<EnhancedKycAsyncResponse, Error>
 
+    /// Fetches the status of a Job. This can be used to check if a Job is complete, and if so,
+    /// whether it was successful.
+    func getJobStatus<T: JobResult>(
+        request: JobStatusRequest
+    ) -> AnyPublisher<JobStatusResponse<T>, Error>
+
+    /// Returns supported products and metadata
+    func getServices() -> AnyPublisher<ServicesResponse, Error>
+
     /// Gets supported documents and metadata for Document Verification
-    /// - Parameter request: request description
-    /// - Returns: description
     func getValidDocuments(
         request: ProductsConfigRequest
     ) -> AnyPublisher<ValidDocumentsResponse, Error>
+
+    /// Returns the different modes of getting the BVN OTP, either via sms or email
+    func requestBvnTotpMode(request: BvnTotpRequest) -> AnyPublisher<BvnTotpResponse, Error>
+
+    /// Returns the BVN OTP via the selected mode
+    func requestBvnOtp(request: BvnTotpModeRequest) -> AnyPublisher<BvnTotpModeResponse, Error>
+
+    /// Submits the BVN OTP for verification
+    func submitBvnOtp(request: SubmitBvnTotpRequest) -> AnyPublisher<SubmitBvnTotpResponse, Error>
 }
 
 extension SmileIDServiceable {
@@ -81,10 +98,6 @@ extension SmileIDServiceable {
 }
 
 public class SmileIDService: SmileIDServiceable, ServiceRunnable {
-    public func getServices() -> AnyPublisher<ServicesResponse, Error> {
-        get(to: "services")
-    }
-
     @Injected var serviceClient: RestServiceClient
     typealias PathType = String
 
@@ -102,21 +115,43 @@ public class SmileIDService: SmileIDServiceable, ServiceRunnable {
         upload(data: zip, to: url, with: .put)
     }
 
-    public func getJobStatus<T>(
-        request: JobStatusRequest
-    ) -> AnyPublisher<JobStatusResponse<T>, Error> {
-        post(to: "job_status", with: request)
-    }
-
     public func doEnhancedKycAsync(
         request: EnhancedKycRequest
     ) -> AnyPublisher<EnhancedKycAsyncResponse, Error> {
         post(to: "async_id_verification", with: request)
     }
 
+    public func getJobStatus<T>(
+        request: JobStatusRequest
+    ) -> AnyPublisher<JobStatusResponse<T>, Error> {
+        post(to: "job_status", with: request)
+    }
+
+    public func getServices() -> AnyPublisher<ServicesResponse, Error> {
+        get(to: "services")
+    }
+
     public func getValidDocuments(
         request: ProductsConfigRequest
     ) -> AnyPublisher<ValidDocumentsResponse, Error> {
         post(to: "valid_documents", with: request)
+    }
+
+    public func requestBvnTotpMode(
+        request: BvnTotpRequest
+    ) -> AnyPublisher<BvnTotpResponse, Error> {
+        post(to: "totp_consent", with: request)
+    }
+
+    public func requestBvnOtp(
+        request: BvnTotpModeRequest
+    ) -> AnyPublisher<BvnTotpModeResponse, Error> {
+        post(to: "totp_consent/mode", with: request)
+    }
+
+    public func submitBvnOtp(
+        request: SubmitBvnTotpRequest
+    ) -> AnyPublisher<SubmitBvnTotpResponse, Error> {
+        post(to: "totp_consent/otp", with: request)
     }
 }